requirements.txt
kafka-python==2.0.1
elasticsearch==7.8.1
elastic-apm==6.0.0
requests==2.22.0
pytest==5.3.5
